career-recommender/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
spacy
sentence-transformers
python-docx
pdfminer.six
tqdm
joblib
transformers
torch
